requirements.txt
Flask==2.1.2
Flask-JSON==0.3.4
pycryptodome==3.15.0
Flask_CORS==3.0.10
